00.00.006
Finished the follow commands:
- SBIT (set bit "output")
- CBIT (clear bit "output")
- GBIT (get bit "output")
- SOUT (set output port)
- GOUT (get output port)
- GINP (get input port)

Almost all commands are verfied and tested at 90%. Only the "GINP" I could not test, because I have not all what I need on the hardware; means by sending this command you receive always the same value, because I can't switch the inputs.
Anyway, these commands should be validated again with a proper equipment.

To write less code I programmed the subroutines "funct_OutputHandler" and "funct_IOhandler". Inside the subroutine "funct_IOhandler" the "_GetBit" for the inputs is not programmed because this will be a new command for JohnsonElectric. I finish this point during the next update but you still can use the actually code. This has no influence on the application!