00.00.008
I changed every command which has not parameter that we send with, for example:
- GOUT
- GINP
- GVER

The problem was you could send GVER,0,100,20 and you received back an E<CR>. Now I changed that for that the customer receive an error, that says "Number of received parameters do not match with used commands".

All commands were tested and verfied.