FW version: 00.00.009
I changed the commands:
- SRACC
- SRDEC

The part with the comma check "0xFFFFFFFF" does not really work like JE would at the beginning. I deleted this from the subroutine because this makes not sence to use. Ask me if you need more informations. Now the customer has always to send a pair of parameters (number of steps with a frequency).

I also renamed the array for the bipolar micro stepping current levels from uint16_CurrentLevel to uint16_BipILevel.