I2C - first steps
All seems to working well but only for the read/receive part there is an issue on the ACK/NACK and stop condition. I don't know how to wait between this both things, because I send at the moment both on the same time. So the stop condition does not work.