UART code correction
The switch case inside the subroutine "uart_set" was wrong for the parity definition.
There was written the follow:
case _NONE: U2MODEbits.PDSEL = 3;

But 3 stands for 9 data-bit no parity. But for 8 data bit with none parity it must be set to 0. Now it is working correct for both UARTs.