Integration of the switch on/off function
Use the SMTYP or RUN command to switch on the output/Mosfet.

Commands: SMCRSTP, GMCRSTP, SPHC and GPHC deleted because we don't need them. I let the code, if once we have to re-use it.

Main:
Switch on of supply deleted because now we can use the SMTYP for that. The "oBiResetSignal" and "oBiPhCurrCtrl" copied into the function "InitA3981".