Working on the ADS1115
Integrated a first version of the ADS1115 in our source code. It seems to work but I have to do all the test's with a real situation. To use the ADS1115 enable it in the periph file. The subroutines are already inside there but in commentary like we haven't the ADS1115 on board.