Working on the ramps issue
It seems that the issue comes from the subroutine "funct_ProtDev". If this routine is active then the bipolar and unipolar are not working anymore. At the moment I put it in commentary and will try to find out the reason later. The bipolar and unipolar source code is now according to the old code from the unipolar driver project. This code has a better protection for acc, run and dec.